+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "third_party/cmocka"]
 	path = third_party/cmocka
 	url = https://github.com/AutoGentoo/cmocka.git
+[submodule "third_party/reflex"]
+	path = third_party/reflex
+	url = git@github.com:AutoGentoo/RE-flex.git
