+++ cmake/test.cmake
@@ -1,5 +1,5 @@
 function(setup_cmocka)
-    include(${PROJECT_SOURCE_DIR}/third_party/cmocka/cmake/Modules/AddCMockaTest.cmake)
+    include(${PROJECT_SOURCE_DIR}/third_party/cmocka/AddCMockaTest.cmake)
     include(${PROJECT_SOURCE_DIR}/cmake/AddMockedTest.cmake)
 endfunction()
