+++ src/codegen/bootstrap/grammar.h
@@ -58,7 +58,7 @@ enum
 };
 
 typedef union {
-    struct KeyVal* key_val;
+    KeyVal* key_val;
     struct LexerRuleProto* l_rule;
     struct Token* token;
     struct GrammarRuleSingleProto* g_single_rule;
